- remove database store from UserRepository.kt
