+++ firebase-api-library/build.gradle
@@ -51,7 +51,7 @@ dependencies {
 
     implementation 'com.google.firebase:firebase-core:18.0.0'           //  FireBase core
     implementation 'com.google.firebase:firebase-firestore:22.0.1'      // FireBase FireStore
-    implementation 'com.google.firebase:firebase-auth:20.0.1'           //  FireBase authentication
+    implementation 'com.google.firebase:firebase-auth-ktx:20.0.1'       //  FireBase authentication
     implementation 'com.google.firebase:firebase-analytics:18.0.0'      //  FireBase analytics
     implementation 'com.google.firebase:firebase-crashlytics:17.3.0'    //  FireBase crashlytics
     implementation 'com.google.firebase:firebase-firestore:22.0.1'      //  FireBase FireStore
